+++ resources/lang/cs/harvest.php
@@ -6,6 +6,7 @@ return [
         'started' => 'Zahájeno',
         'progress' => 'Zpracováva se (:current/:total)',
         'error' => ':error',
+        'finished' => 'Zpracovaných záznamů :processed',
         'completed' =>
             'Zpracovaných záznamů: :processed' . PHP_EOL .
             ':created nových' . PHP_EOL .